Renamed storybook pages
